Updated USI_NF_Mode to play nicer with SETI
SETI/rebalance/ctt now appears to primarily modify costs and some mass for existing (and a handful of reskinned) props, but it doesn't do anything with power consumption.  The SETI electrics section primarily deals with TweakScale values and cost.  We probably shouldn't mess with electrical values unless something like NearFutureElectrical adds more parts.

In addition it looks like the SETI module manager flag is depreciated so it shouldn't be used anyway.  The only thing that still sets it is the SETI community tech tree and that just moves things around.